+++ vira.hs
@@ -10,6 +10,7 @@
         ]
      , build.flakes =
          [ "./examples/flake-parts" { overrideInputs = cf }
+         , "./examples/standalone" { overrideInputs = cf }
          , "./test/flake-parts" { overrideInputs = cf }
          , "./test/flake-parts-no-defaults" { overrideInputs = cf }
          , "./test/flake-parts-override-default" { overrideInputs = cf }